ARM64: preserve 64-bit 64DD DMA pointers

+++ src/libultrare/io/devmgr.c
@@ -30,7 +30,7 @@ void __osDevMgrMain(void *args)
       info->sectorNum = -1;
       if (info->transferMode != LEO_SECTOR_MODE)
       {
-        blockInfo->dramAddr = (void *)((u32)blockInfo->dramAddr - blockInfo->sectorSize);
+        blockInfo->dramAddr = (void *)((uintptr_t)blockInfo->dramAddr - blockInfo->sectorSize);
       }
       if (info->transferMode == LEO_TRACK_MODE && mb->piHandle->transferInfo.cmdType == LEO_CMD_TYPE_0)
         messageSend = 1;
